backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "langchain>=0.1.0",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.7",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
